breh moment (#142)
* command and other minor tweaks

* fix conflict cuz offz push to master and not dev so dev is outdated bruh moment

* space of

* Guild Chat (#120)

* guildchat?

* fixes, remove admin cmds add player cmds

* move shit

* fix missing imports

* cleanup

* tab completion (#122)

* tab completion

* fix sub commands sort of

* finish tabcomplete

* fix guildchat not taking priority over venturechat

* dumb mistakes

* birdswatter, baneofkuon, jawbreaker (#125)

* Fix for GSit bypassing curse (#129)

* fix curse bypass for gsit

* fix curse and flying machines & move function to anticheese

* remove hacky and broken curse-check

* gsitlistener

* make handleCurse helper func and move gsit related events

* gsit on le repo
