+++ build.gradle.kts
@@ -20,6 +20,7 @@ allprojects {
         maven("https://maven.pkg.jetbrains.space/public/p/compose/dev")
         maven("https://jitpack.io")
         maven("https://repo.codemc.io/repository/maven-snapshots/")
+        maven("https://repo.mineinabyss.com/releases")
     }
 
     dependencies {
@@ -47,6 +48,7 @@ allprojects {
         compileOnly("com.mineinabyss:deeperworld:$deeperWorldVersion")
         compileOnly("com.github.MilkBowl:VaultAPI:1.7") { exclude(group = "org.bukkit") }
         compileOnly("nl.rutgerkok:blocklocker:1.10.2-SNAPSHOT")
+        compileOnly("com.gecolay:gsit:1.0.6")
 
         implementation("com.mineinabyss:idofront:$idofrontVersion")
     }
