+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/anticheese/AntiCheeseFeature.kt
@@ -1,6 +1,7 @@
 package com.mineinabyss.anticheese
 
 import com.mineinabyss.idofront.plugin.registerEvents
+import com.mineinabyss.mineinabyss.core.AbyssContext
 import com.mineinabyss.mineinabyss.core.AbyssFeature
 import com.mineinabyss.mineinabyss.core.MineInAbyssPlugin
 import kotlinx.serialization.SerialName
@@ -10,6 +11,7 @@ import kotlinx.serialization.Serializable
 @SerialName("anticheese")
 class AntiCheeseFeature : AbyssFeature {
     override fun MineInAbyssPlugin.enableFeature() {
+        if (AbyssContext.isGSitLoaded) registerEvents(GSitListener())
         registerEvents(AntiCheeseListener())
     }
 }
