+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/anticheese/AntiCheeseListener.kt
@@ -1,11 +1,15 @@
 package com.mineinabyss.anticheese
 
+import com.mineinabyss.helpers.handleCurse
 import com.mineinabyss.idofront.messaging.error
 import com.mineinabyss.mineinabyss.core.layer
+import dev.geco.gsit.api.GSitAPI
+import dev.geco.gsit.api.event.PlayerGetUpSitEvent
 import org.bukkit.ChatColor
 import org.bukkit.entity.Player
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
+import org.bukkit.event.block.BlockPistonExtendEvent
 import org.bukkit.event.block.BlockPlaceEvent
 import org.bukkit.event.entity.EntityPotionEffectEvent
 import org.bukkit.potion.PotionEffectType
@@ -35,5 +39,20 @@ class AntiCheeseListener: Listener {
                 player.error("${ChatColor.BOLD}Slow Falling ${ChatColor.RED}has been disabled")
             }
         }
+    }
+}
+
+class GSitListener : Listener {
+
+    // Cancels pistons if a player is riding it via a GSit Seat
+    @EventHandler
+    fun BlockPistonExtendEvent.seatMovedByPiston() {
+        if (GSitAPI.getSeats(blocks).isEmpty()) return
+        else isCancelled = true
+    }
+
+    @EventHandler
+    fun PlayerGetUpSitEvent.handleCurseOnSitting() {
+        handleCurse(player, seat.location.toBlockLocation(), player.location)
     }
 }
